refactor: split REMD tests
